requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
python-multipart>=0.0.5
kokoro>=0.8.2
misaki[zh]>=0.8.2
soundfile
torch>=1.10.0
pytest>=6.2.5
